Fix BukkitConfigMappers

+++ MrCore-Bukkit/src/main/java/me/mrletsplay/mrcore/bukkitimpl/config/BukkitConfigMappers.java
@@ -59,8 +59,8 @@ public class BukkitConfigMappers {
 						json.getDouble("x"),
 						json.getDouble("y"),
 						json.getDouble("z"),
-						json.has("yaw") ? (float) json.getDouble("yaw") : 0f,
-						json.has("pitch") ? (float) json.getDouble("pitch") : 0f
+						json.has("yaw") ? json.getFloat("yaw") : 0f,
+						json.has("pitch") ? json.getFloat("pitch") : 0f
 					)
 			)
 			.mapString("world", l -> l.getWorld().getName(), null).then()
